Move crud.js into a modules folder for better modularity
